refactor: packages separated from the component
