+++ src/app/photographer/quotation/page.tsx
@@ -1,3 +1,4 @@
+import { PackageForm } from '@/components/package-page/create-package'
 import PhotographerQuotation from '@/components/quotation/photographer-quotation'
 import { Button } from '@/components/ui/button'
 
@@ -27,6 +28,24 @@ const quotations: quotation[] = [
   },
 ]
 
+const packages: PackageForm[] = [
+  {
+    name: 'Wedding Package',
+    packageDescription: 'Wedding photography package',
+    price: 400,
+  },
+  {
+    name: 'Birthday Package',
+    packageDescription: 'Birthday photography package',
+    price: 300,
+  },
+  {
+    name: 'Graduation Package',
+    packageDescription: 'Graduation photography package',
+    price: 200,
+  },
+]
+
 export default function QuotationPage() {
   return (
     <div className='flex w-full flex-col gap-6 px-4 py-4 lg:px-32'>
@@ -36,7 +55,7 @@ export default function QuotationPage() {
         <Button>New Quotation</Button>
       </div>
 
-      <PhotographerQuotation quotations={quotations} />
+      <PhotographerQuotation quotations={quotations} packages={packages} />
     </div>
   )
 }
